combine gcoin_test and test
