--- src/gcoin_test/Makefile
@@ -1,6 +1,0 @@
-all:
-	$(MAKE) -C .. gcoin_test
-clean:
-	$(MAKE) -C .. gcoin_test_clean
-check:
-	$(MAKE) -C .. gcoin_test_check
