+++ src/test/rpc_wallet_tests.cpp
@@ -17,7 +17,7 @@
 using namespace std;
 using namespace json_spirit;
 
-extern Array createArgs(int nRequired, const char* address1 = NULL, const char* address2 = NULL);
+extern Array createArgs(int nRequired, const char* address1, const char* address2);
 extern Value CallRPC(string args);
 
 extern CWallet* pwalletMain;
